Allow user to add a new artist to the deck
The function addNewArtist allows the user to add a new musical artist to the deck.  It sends a POST request to the server with the user input from NewUserForm.  Upon a successful request, the function sets state back from the form to the deck, and sets the added card as the current card so the user can see their created card in the deck.